+++ src/index.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { Provider } from 'react-redux';
 import ReactDOM from 'react-dom';
 import './index.css';
-import store from './store/reducers';
+import store from './redux/reducers';
 import Routes from './Routes';
 
 ReactDOM.render(
